# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/users/readings:
    get:
      tags:
      - api-statistics-controller
      summary: Get users that have read the most books
      operationId: getUsersTotalReadings
      responses:
        "200":
          description: Users found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_Username'
        "204":
          description: No content
          content:
            '*/*':
              schema:
                type: object
...
  /api/users/all:
    get:
      tags:
      - api-statistics-controller
      summary: Get the total number of users
      operationId: getTotalUsers
      parameters:
      - name: count
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Total users found
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /api/genres:
    get:
      tags:
      - api-statistics-controller
      summary: "Get the most read genres (if count is true, return the count of genres)"
      operationId: getMostReadGenresGeneral
      parameters:
      - name: count
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Most read genres found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre_BasicInfo'
        "204":
          description: No content
          content:
            '*/*':
              schema:
                type: object
  /api/genres/me:
    get:
      tags:
      - api-statistics-controller
      summary: Get the most read genres for the logged user
      operationId: getMostReadGenresUser
      responses:
        "200":
          description: Most read genres found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre_BasicInfo'
        "204":
          description: No content
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
  /api/genres/all:
    get:
      tags:
      - api-statistics-controller
      summary: Get the total number of genres
      operationId: getTotalGenres
      parameters:
      - name: count
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Total genres found
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
  /api/books/me/recommended:
    get:
      tags:
      - api-statistics-controller
      summary: Get recommended books for the logged user
      operationId: recommendedBooks
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      - name: by
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Recommended books found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_GenreBookBasicView'
        "204":
          description: No content
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
  /api/books/all:
    get:
      tags:
      - api-statistics-controller
      summary: Get the total number of books
      operationId: getTotalBooks
      parameters:
      - name: count
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Total books found
          content:
            application/json:
              schema:
                type: integer
                format: int64
  /api/authors:
    get:
      tags:
      - api-statistics-controller
      summary: "Get the most read authors (if count is true, return the count of authors)"
      operationId: getMostReadAuthorsGeneral
      parameters:
      - name: count
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "204":
          description: No content
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Most read authors found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author_BasicInfo'
  /api/authors/me:
    get:
      tags:
      - api-statistics-controller
      summary: Get the most read authors for the logged user
      operationId: getMostReadAuthorsUser
      responses:
        "204":
          description: No content
          content:
            '*/*':
              schema:
                type: object
        "200":
          description: Most read authors found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author_BasicInfo'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
  /api/authors/all:
    get:
      tags:
      - api-statistics-controller
      summary: Get the total number of authors
      operationId: getTotalAuthors
      parameters:
      - name: count
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: Total authors found
          content:
            application/json:
              schema:
                type: integer
                format: int64
...
components:
  schemas:
...
    Book_Username:
      type: object
      properties:
        id:
          type: integer
          format: int64
        isbn:
          type: string
    User_Username:
      type: object
      properties:
        username:
          type: string
        readBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_Username'
        readingBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_Username'
        wantedBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_Username'
        role:
          type: array
          items:
            type: string
...
    Genre_BasicInfo:
      type: object
      properties:
        name:
          type: string
    Book_GenreBookBasicView:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        title:
          type: string
        authorString:
          type: string
        description:
          type: string
        releaseDate:
          type: string
        ISBN:
          type: string
        averageRating:
          type: number
          format: double
        genre:
          $ref: '#/components/schemas/Genre_GenreBookBasicView'
        series:
          type: string
        pageCount:
          type: integer
          format: int32
        publisher:
          type: string
        id:
          type: integer
          format: int64
        isbn:
          type: string
    Genre_GenreBookBasicView:
      type: object
      properties:
        name:
          type: string
    Author_BasicInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string